Introduce a CommandProcessor class.
Instances of this class are supposed to perform the actual
processing of commands. The method ``process`` takes an iterable
of lines and tries to interpret them as commands.

This way, when processing commands received by email, we only
extract the commands from the message and pass them on to a
CommandProcessor instance to get the result which is emailed back
to the users.