Fix the statistics management command.
The statistics management command claims to output the number of
source packages with at least one subscription, but it was actually
the number of all packages (including pseudo- and subscription only).

Added a new helper member to the package model manager to retrieve
the QuerySet of packages with at least one subscriber.