Add class which updates locally cached Debian repository data.
This class utilizes the python-apt package to build a local cache
of source and binary package entries for the Repository model
instances that are registered.

It exposes a method to update the cache and returns a list of updated
files. Clients can use this information for their own purposes, such as
updating the database information for updated packages.