Handle failed tasks.
Tasks which raise an exception during their execution now do not
halt all subsequent (possibly independent) tasks.
A failed task simply does not emit any events which causes the tasks
that are dependent on it not to run.
For now, all tasks are responisble to leave the database in a
consistent state themselves -- the job executor does no transaction
management.