Add a BaseTask subclass with common operations for Package update tasks.
For now, this class only handles settings the force_update parameter.

With this, a forced update of packages based on apt cache data can be
triggered by passing the force_update argument to the run_task function
when starting a RepositoryUpdateTask.
The forced update means that the apt cache will be cleared and
re-downloaded so that all further processing is based on this newly
retrieved repository data.