Make the binaries panel scrollable after 10 items.
After the binaries panel reaches more than 10 items, a scrollbar
appears.
The scrollbar is implemented using the perfect-scrollbar jQuery
plugin for desktop to allow a stylized scrollbar in a
cross-browser compatible way.
On mobile we let the browser handle the scrolling to make sure it
works correctly for touch gestures as well.