Add a minimum height option for the scrollbar drag.
The perfect-scroll plugin didn't have a lower limit on the drag size
so in containers significantly smaller than their content, the drag
size would be miniscule. (e.g. linux package binaries)

This adds a new option which adds a lower limit to the drag size so
it is always clearly noticeable when a container is scrollable.