Implement extracting lzma compressed tarballs in Python2.
Python2's tarfile package does not support lzma (xz) compressed files
automatically. It is necessary to manually uncompress the file using
the lzma module and pass this to tarfile to extract the files in the tar.

In Python3, this is left up to tarfile.

The docs were also updated to indicate a new package dependency for
Python2.7.