Refactor UpdateRepositoriesTask method for marking non-updated files.
When a file is not updated by the apt cache, all packages found in it
must be kept unchanged after the update. The method marks them all as
packages which should be kept.

After the refactoring, the method receives a manager instance which
handles the correct package repository entries (either source or
binary entries, based on which file type is passed). This way, the
same method can handle both binary and source package files.