Implement user log in.
Users are allowed to log in using any of their associated emails. A
custom authentication backend is implemented which matches the provided
email to a User model instance.  The backend's authentication method
signature matches the default Django backend in order to allow admin
users to log in using any of their emails, as well.