Refactor function for creating news items to a EmailNews manager method.
The function was a global function in the pts.mail.main_news module.
A more logical place for such a function is in a model Manager so that
other code can depend on it.

The method is generalized slightly more than the function was and
besides automatically extracting the subject (and using it for the
news title), it allows the user to override that default value
with their own.