Implement account merges.
When a user tries to add an email which already belongs to a
different account, he is offered the option to merge the two accounts.

The merge must be confirmed by following a link provided in an email
sent to the email that does not belong to the account that tried
adding it and is only accessible while logged in with the other
account.