Implement Debian-specific hooks for login/logout.
When active, the log in hook does not allow any user with a
@debian.org email address associated to his account to log in via the
PTS log in form. It throws an error directing the user to sso.d.o.

When active, the log out hook redirects the users to sso.d.o service
which logs them out not only in the PTS level, but also in the SSO
level.