Replace transaction.commit_on_success with transaction.atomic
The former is deprecated and generates test suite failures
with Django 1.6:

  TransactionManagementError: Your database backend doesn't behave
  properly when autocommit is off. Turn it on before using 'atomic'.

This is because commit_on_success disables autocommit and you get
this as soon as you try to start an inner transaction (even implicitly
with a model's get_or_create() call).

For compatibility with Django 1.5, alias transaction.atomic to
transaction.commit_on_success.