Fix a test-suite failure with Django 1.6.2 and newer
Until version 1.6.1, Django was generating mails with the as_string()
method on the object returned by the message() method of a
django.core.mail.EmailMessage object. To avoid problemes in Python 3, we
used to override the default as_string() to return bytes in Python3 too.
This was the purpose of BytesEmailMessage.

Since version 1.6.2, it's now using the as_bytes() method which doesn't
exist in Python 2 but that they added in their own SafeMIMEText class.

Implementing as_bytes() is BytesEmailMessage could be enough to fix the
issue but I found it cleaner to get rid of this class and to monkey patch
the email.Message objects generated by our own message_from_bytes()
to override the as_string/as_bytes methods so that they behave like
Django expects them to behave and so that we ensure identical behaviour
between Python 2 and 3 with older versions of Django too.