Refactor the way settings are defined
Settings with default values computed from other settings are now defined
in a last step and only if they have not been defined before. This make it
possible to change only the most important settings and have the dependent
setting automatically reflect the change on that important setting.

Simplify the default debian configuration to benefit from that.