Fix public links to .dsc and to pool directory
When public_uri is set, it should be preferred over uri since it contains
an URI that is usable by anyone while uri references an URI that probably
works only in the context of the distro-tracker server (i.e. a file:// URI
or an http:// server on a private network).