settings: drop local.py, instead provide local.py.sample
Always having uncommited changes that we don't want to commit on local.py
is a major pain. Restructure the logic so that we provide a
local.py.sample that users should copy to local.py. If that file
is unavailable, then we fallback to selected.py.

Also move SQLite database settings in a dedicated file so that we
can force usage of that database by re-importing the DATABASES
setting from that module.