core: fix test PackageViewTest.test_catch_all_redirect
Accessing /admin/ with Django 1.7 will redirect the unauthenticated
user to the login screen so the lack of redirect is not a good test
to verify that the URL is not handled by the usual package redirect.
Instead rely on the fact that the admin package doesn't exist and
that we don't get a 404 that we would get if we were redirected
to /pkg/admin.