core: drop intermediary class from registered panels
The version of six.with_metaclass() bundled with Django 1.6 creates an
intermediary class that gets registered in the list of panels and then
outputs a useless None.

The version of six in Django 1.7 has a different behaviour where the
intermediary class doesn't get registered and thus doesn't have the
unregister_plugin class method either.

Based-on-patch-by: Christophe Siraut <d@tobald.eu.org>
Closes: #756269