accounts: fix is_subscribed_to()'s logic
The unit tests were actually passing but the method was doing the wrong
thing nevertheless. A functional test failure led me to this problem.

package.subscriptions is a list of EmailSettings objects and no longer
of UserEmail so we need an indirection in the .filter() to retrieve
the associated UserEmail objects.