Add infrastructure to generate coverage report
bin/gen-coverage.sh is the script that you should run to generate
a coverage report in the "coverage.html" directory. It will use
tox to run the tests in a Python 2 based environment and then a Python 3
based environment and will combine the result of both run.

There's some hackery needed to make it work with both the Debian packages
python{3,}-coverage and the pypi-installed coverage. The former are using
python-coverage and python3-coverage as executables while the latter uses
the (much too generic) "coverage".