core/models: tweak the logic to return a short description
When we have a single binary package associated to the source package, it
has a good description by definition.

However when we have multiple binary packages, picking one at random is
not a good idea. We can rely on the description of the binary package
of the same name but nothing more.

Reported-by: Ben Hutchings