mail: New MailProcessor class
This class takes a message as input and knows how to identify
the Distro Tracker service that should handle the given message.
When the process() method is called, it does its analysis and
hands out control the respective service for further processing.

Distro Tracker provides 3 services by default: control, dispatch
and bounces.