mail: handle bounces received when original case has been changed
There are a bunch of MTA that lowercase the email in the Delivered-To
header and as such we need to be able to find the user email in a
case-insensitive way.

Also we should be ready to handle the case when we get a bounce for
a user that no longer exists for some reason.