test: refactor handling of keyring directory
The keyring directory is already part of the standard temporary
directories that are created and cleaned. Thus move the ignore_errors
attribute required by the conccurent removal of gpg-agent socket
to the shutil.rmtree call handling the standard temporary directories.

Get rid of the empty "TEST_KEYRING_DIRECTORY" since the standard
settings.DISTRO_TRACKER_KEYRING_DIRECTORY is also empty by default.

Do not use the GNUPGHOME environment variable but instead pass the
directory as parameter to the Context() object.