Fix UserAuthMixin to behave deterministically
When you had multiple users, the login() method without any username
specified would use a random user. Now it will raise an exception. You
have to specify the username in that situation.

Also update the code in a way that doesn't modify the original
UserAuthMixin.USERS, instead we fully override it every time. However
to avoid too much duplication, we have improved it to attribute default
email and passwords.