UpdateExcusesTask: tweaks the generation of excuses list to add missing entries
The YAML file doesn't provide a statement on the age of the package in unstable
in the packages' excuses entry. It doesn't provide anything on blocking items
or on the migration status either. Also add the result of is-candidate as a
final verdict in the end of the excuses list.