settings: use lists instead of tuples for many settings
All the modified settings default to lists in the current version of
Django. Special care must be taken to ensure that we duplicate the list
objects in the various namespaces otherwise we get a nice mess with a
single shared configuration combining all the operations of all the
settings that are loaded but not necessarily imported in the main settings
namespace.

We add __all__ attributes to some of the core modules to better control
what settings are actually exported.

This also fixes the regression introduced in
dac264a2caabbd45f7a483a2c678bad443eb6162 where the debug_toolbar
application suddenly started to be required for tests too.