Do not choke on maintainer emails without a plus sign
The code tries to parse the localpart as "team+foo" when the email is
hosted in the domain dedicated to distro-tracker. But it failed badly
when the maintainer email did not contain any plus sign.

This commit fixes the following exception:

Traceback (most recent call last):
  File "/home/rhertzog/deb/core/distro-tracker/distro_tracker/core/tests/tests_retrieve_data.py", line 1185, in test_with_unqualified_team_email
    self.run_task()
  File "/home/rhertzog/deb/core/distro-tracker/distro_tracker/core/tests/tests_retrieve_data.py", line 1109, in run_task
    self.task.execute()
  File "/home/rhertzog/deb/core/distro-tracker/distro_tracker/core/retrieve_data.py", line 1094, in execute
    self.add_package_to_maintainer_teams(package, maintainer)
  File "/home/rhertzog/deb/core/distro-tracker/distro_tracker/core/retrieve_data.py", line 1060, in add_package_to_maintainer_teams
    service, slug = localpart.split('+', 1)
ValueError: not enough values to unpack (expected 2, got 1)