refactoring: Move the logic of bugs displaying from rules to a new class
It introduces a new approach to handle the logic related to the
presentation of bugs data through a new management class. Vendors may
extend this new class to customize the way bugs information are
presented in tables and panels.
This commit also changes the debian app to reuse the debian's panel
content inside bugs table field.