tasks: force usage of string for the keys in ProcessModel
After a save/reload in a JSON dict, the integer keys will be converted
to a string anyway. So we have to cope with this.

Use this opportunity to reduce the amount of data retrieved from the
database in ProcessModel tasks by filtering out the unwanted entries in
the SQL query directly.