tests: fix functional tests
The new test was working fine standalone but was not working
when run after any other test. This is due to Django deleting all
data from the database after each test in TransactionTestCase
and LiveServerTestCase is based on it.

To work-around this and ensuring that we have the desired
initial data, we have to set "serialize_rollback = True"
on the TransactionTestCase. However this fails unless I exclude
the contenttypes application from this serialization.