Enable flake8 logging-format plugin
The plugin is not very smart. It applies its analysis to any method
whose names matchs one of the usual logging methods. Thus it had lots
of false positives that I had to work-around by refactoring the existing
code. In the end, it's still a good idea to make any error/warning
method have a signature that matches the one used by the logging module.