django_email_accounts: validate email on UserEmail.save()
The goal is to never accept invalid data in the database. It will
raise ValidationError but any code creating UserEmail should either
check emails validity before or be ready to handle those errors.

The validation also applies to the user field, so that's why we must
add the blank=True, otherwise validation does not pass.