debian: better handle 'dak rm' mails
The code processing 'dak rm' mails was working on the assumption that we
get a single mail for all the packages listed in the mail. That
assumption is no longer true as we get a copy for each source package.
We should thus generate a news only for the specific package that was
targetted by the current email otherwise we store multiple news on each
source package listed in the mail.

Closes: #962000