mail: update tracker_receive_news to rely on classify_message
Right now there's two code paths to inject a mail as a news item.
The code path embedded in classify_message has been added afterwards
so that we can get all the desired information out of the usual flow of
incoming messages.

The older path is unused except for this management command. So we
update the management command to call classify_message() and benefit
from the side effect of having news items created.