Add license metadata to JavaScript files
This format should enable LibreJS to recognize that the
JavaScript is free. Changes were made by hand verifying
the licenses of the upstream projects.

In particular jquery.cookie.js 1.3.0 had switched from
being dual-licensed to mono-licensed, but it appears this
wasn't reflected in the license header, so I revised it.