Work around two failures in functional tests
In RepositoryAdminTest we need to wait a bit longer otherwise clicking
the Repositories link might not work as expected.

For ManageSubscriptionTest the "serialization_rollback" feature does not
seem to work properly, we get a FOREIGN KEY error during that rollback.
So drop this and instead recreate the desired Keyword objects during
setUp().