core: change default structure of the HttpCache
The HttpCache was not friendly with administrators because it was not
easy to inspect for a human. The giant directory with thousands of files
based on md5sums was also not very efficient.

Change the logic to use a directory tree based on the URLs paths. It
makes it easy to drop specific parts of the cache and it makes it easier
to inspect the content of the files fetched by distro-tracker.