Handle security issues according to each release status (1)
From a security point of view, the release status is defined in
https://security-tracker.debian.org/tracker/distributions.json
through the "support" key.

When a release is marked "end-of-life", we don't want to show anything.
When a release is not marked "security", we don't want to show the
special instructions provided by the security team.

This commit implements the first step towards this: retrieve the
information from the security tracker and make it available for further
processing.