Handle security issues according to each release status (5)
The release status can change at any time while the list of CVE
stays unchanged and we still want to update the various action items
in that case.

We now monitor the content of distributions.json and set
force_update=True when we detect that it has changed so that
all action items are processed even if the underlying CVE list
is unchanged.

With this last commit, we can consider the patchset to be complete.

Fixes #58