debian: drop the outdated clang service
clang.d.n does not rebuild (more or less quickly) new versions uploaded
to unstable (so the results are mostly stale), and does not even
provide successful build logs of old mass-rebuilds (so most of the
packages get an empty table).

Also, a potential replacement for it was proposed few years ago:
https://salsa.debian.org/qa/distro-tracker/-/merge_requests/47
however, it never got any updates after the intial bootstrap.

Hence, simply drop the clang link for now, as it brings no value to
users.