Ensure we are able to extract the signature from the message body
When we feed the content of all the parts (including the signature
due to PGP/MIME), we confuse GPG and it will not report any signature
that we want to identify in the body. So we scan only the first
text/plain part of the message to identify the signature.